+++ jaguar_jwt/CHANGELOG.md
@@ -9,6 +9,9 @@
 + Added general support for non-registered claims.
 + Tidy up for static analysis and Dart linter.
 + Implemented toString method for JwtClaim.
++ Allow for customized checking of the JWT header.
++ Fixed use of _splayify/_spaly in toJson and changed dynamic to Object.
++ Improved format of output produced by JwtClaim.toString().
 
 ## 2.1.2